Rewrite admin service by using multiple server feature
